% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGadjust_ratefactor.R
\docType{data}
\name{allfacts}
\alias{allfacts}
\title{Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object}
\format{An object of class \code{character} of length 10.}
\usage{
allfacts
}
\arguments{
\item{df}{dataframe in which rows are populations, column 'factor_df' is a nested dataframe of rate-factors, and column 'pop_prods' is the rowProducts of 'factor_df'.}

\item{pop}{tidyselect variable indicating population ID}

\item{factrs}{character vector of rate-factors}
}
\description{
Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object
}
\examples{
......
}
\keyword{datasets}
