% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasGupt_2pop.R
\name{DasGupt_2pop}
\alias{DasGupt_2pop}
\title{Standardisation and decomposition of rates over P rate-factors and 2 populations.
Outputs a ? object}
\usage{
DasGupt_2pop(df, pop, factrs)
}
\arguments{
\item{df}{dataframe or tibble object, with columns specifying 1) population and 2) each rate-factor to be considered}

\item{pop}{tidyselect variable indicating population ID}

\item{factrs}{character vector of rate-factors}
}
\description{
Standardisation and decomposition of rates over P rate-factors and 2 populations.
Outputs a ? object
}
\examples{
......
}
