% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasGupt_Npop.R
\name{DasGupt_Npop}
\alias{DasGupt_Npop}
\title{Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object}
\usage{
DasGupt_Npop(
  df,
  pop,
  ...,
  baseline = NULL,
  id_vars = NULL,
  ratefunction = NULL
)
}
\arguments{
\item{df}{dataframe or tibble object, with columns specifying 1) population and 2) each rate-factor to be considered}

\item{pop}{tidyselect variable indicating population ID}

\item{...}{tidyselect variables indicating rate-factors}

\item{baseline}{baseline population to standardise against. if NULL then will do Das Gupta's full N-population standardisation.}

\item{id_vars}{character vector of variable names in dataframe to attach back to output (useful if analysing sub-population specific rates)}

\item{ratefunction}{allows user to define rate as a specific function F of factors. This should be a character string of the r syntax, with the factor names. Defaults to the product (e.g., "a*b").}
}
\description{
Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object
}
\examples{
DasGupt_Npop(df,year,prev,freq,baseline=NULL,id_vars=c("age","sex"),ratefunction="prev*freq")
}
