% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasGupt_rates.R
\name{DasGupt_rates}
\alias{DasGupt_rates}
\title{Das Gupta extract adjusted rates}
\usage{
DasGupt_rates(df)
}
\arguments{
\item{df}{DG output}
}
\description{
Das Gupta extract adjusted rates
}
\examples{
......
}
