% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasGupt_table.R
\name{DasGupt_table}
\alias{DasGupt_table}
\title{Das Gupta extract adjusted rates after}
\usage{
DasGupt_table(df, pop1, pop2)
}
\arguments{
\item{df}{DG output}

\item{pop1}{population 1 (string or numeric)}

\item{pop2}{population 2 (string or numeric)}
}
\description{
Das Gupta extract adjusted rates after
}
\examples{
......
}
