% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg354.R
\name{dg354}
\alias{dg354}
\title{Inner function called by dg2pop}
\usage{
dg354(df2, i, factrs, ratefunction, quietly = TRUE)
}
\arguments{
\item{df2}{dataframe in which rows are populations, column 'factor_df' is a nested dataframe of rate-factors, and column 'pop_prods' is the rowProducts of 'factor_df'.}

\item{i}{the index of the factrs vector which is being adjusted for (the alpha in P-alpha)}

\item{factrs}{character vector of rate-factors}

\item{ratefunction}{allows user to define rate as a specific function F of factors. This should be a character string of the r syntax, with the factor names. Defaults to the product (e.g., "a*b").}
}
\description{
Inner function called by dg2pop
}
\examples{
......

}
