% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg611.R
\name{dg611}
\alias{dg611}
\title{Das Gupta equation 6.11: Standardises rates across populations}
\usage{
dg611(srates, all_p, y, factor)
}
\arguments{
\item{srates}{a dataframe/tibble object of standardised rates from \code{\link[=dg2pop]{dg2pop()}}}

\item{all_p}{character or numeric vector of all N populations}

\item{y}{character/numeric indicating a single population}

\item{factor}{string indicating rate-factor being standardised.}
}
\value{
data.frame object including K-a standardised rates for each population for given factor a, across N populations
}
\description{
Das Gupta equation 6.11: Standardises rates across populations
}
