% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DG612numerator.R
\name{DG612numerator}
\alias{DG612numerator}
\title{Very messy logic for DasGuptas equation 6.12
Outputs a vector}
\usage{
DG612numerator(df, y1, y2, yj)
}
\arguments{
\item{df}{dataframe or tibble object, with each columns specifying the factor-effect for all pairwise comparisons of populations.}

\item{y1}{numeric or character vector indicating population 1 of factor-effect A between y1-y2}

\item{y2}{numeric or character vector indicating population 2 of factor-effect A between y1-y2}

\item{y3}{numeric or character vector indicating population j of the DG:6.12 equation (intention is to loop this function taking all other populations as j, thereby standardising y1-y2 factor effect across other populations)}
}
\description{
Very messy logic for DasGuptas equation 6.12
Outputs a vector
}
\examples{
......
}
