% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DG_2pop.R
\name{DG_2pop}
\alias{DG_2pop}
\title{Standardisation and decomposition of rates over P rate-factors and 2 populations.
Outputs a ? object}
\usage{
DG_2pop(df, pop, factrs, ratefunction = NULL)
}
\arguments{
\item{df}{dataframe or tibble object, with columns specifying 1) population and 2) each rate-factor to be considered}

\item{pop}{tidyselect variable indicating population ID}

\item{factrs}{character vector of rate-factors}

\item{ratefunction}{allows user to define rate as a specific function F of factors. This should be a character string of the r syntax, with the factor names. Defaults to the product (e.g., "a*b").}
}
\description{
Standardisation and decomposition of rates over P rate-factors and 2 populations.
Outputs a ? object
}
