% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_check.R
\name{dg_check}
\alias{dg_check}
\title{Performs some initial checks of the input data for dgnpop(), providing messages where issues are flagged}
\usage{
dg_check(x, pop, factors, id_vars = NULL, crossclassified = NULL)
}
\arguments{
\item{x}{dataframe or tibble object, with columns specifying 1) population, 2) each rate-factor to be considered, and (optionally) 3) variables indicating underlying subpopulations}

\item{pop}{name (character string) of variable indicating population}

\item{factors}{names (character vector) of variables indicating compositional factors}

\item{id_vars}{character vector of variables indicating sub-populations}

\item{crossclassified}{character string of variable indicating size of sub-population. If specified, the proportion of each population in a given sub-population (e.g. each age-sex combination) is re-expressed as a product of symmetrical expressions representing the different variables (age, sex) constituting the sub-populations. These expressions are then used as compositional factors in the standardisation. If NULL, then providing a single variable as a compositional factor that represents the proportion of the population in each given sub-population will combine the contribution of all sub-population variables.}
}
\value{
error message processing
}
\description{
Performs some initial checks of the input data for dgnpop(), providing messages where issues are flagged
}
