% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DG_Npops_std.R
\name{DG_Npops_std}
\alias{DG_Npops_std}
\title{Das Gupta equation 6.11: Standardises rates across populations
Outputs a tibble object of the rate factor F in population Y, standardised across all populations}
\usage{
DG_Npops_std(srates, all_p, y, fctr)
}
\arguments{
\item{srates}{a dataframe/tibble object of ?? can't remember!!!}

\item{all_p}{character or numeric vector of all N populations}

\item{y}{character of numeric value indicating a single population}

\item{fctr}{string indicating rate-factor being standardised.}
}
\description{
Das Gupta equation 6.11: Standardises rates across populations
Outputs a tibble object of the rate factor F in population Y, standardised across all populations
}
\examples{
......
}
