% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_plot.R
\name{dg_plot}
\alias{dg_plot}
\title{Creates a plot of Das Gupta standardised rates across the set of populations}
\usage{
dg_plot(
  dgo,
  xlab = "population",
  ylab = "rate",
  legend.position = "topright",
  ...
)
}
\arguments{
\item{dgo}{output from \code{\link[=dgnpop]{dgnpop()}}}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{legend.position}{legend position, passed to \code{\link[graphics:legend]{graphics::legend()}}, choose from "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center", or provide xy.coords}

\item{...}{other arguments passed to methods, such as \code{\link[graphics:par]{graphics::par()}}}
}
\value{
A plot of each of the set of K-a standardised rates across populations
}
\description{
Creates a plot of Das Gupta standardised rates across the set of populations
}
