% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_table.R
\name{dg_table}
\alias{dg_table}
\title{Creates a small table of Das Gupta standardised rates. If no populations are specified, rates will be shown for all available populations. If only two populations (or if two particular populations are specified), then rate-differences and 'decomposition effects' are calculated and presented.}
\usage{
dg_table(dgo, pop1 = NULL, pop2 = NULL)
}
\arguments{
\item{dgo}{output from \code{\link[=dgnpop]{dgnpop()}}}

\item{pop1}{optional name of first population for decomposition (character/numeric)}

\item{pop2}{optional name of second population for decomposition (character/numeric)}
}
\value{
data.frame object with rows for each of the K-a standardised rates and the crude rates, and columns for each of the N populations. When only two populations are included, or if two populations are explicitly specified, standardised rate differences are provided, and are also expressed as a percentage of the crude rate differences (typically referred to as 'decomposition effects').
}
\description{
Creates a small table of Das Gupta standardised rates. If no populations are specified, rates will be shown for all available populations. If only two populations (or if two particular populations are specified), then rate-differences and 'decomposition effects' are calculated and presented.
}
