% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGadjust_ratefactor_old.R
\name{DGadjust_ratefactor_old}
\alias{DGadjust_ratefactor_old}
\title{Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object}
\usage{
DGadjust_ratefactor_old(df2, pop, i, factrs)
}
\arguments{
\item{pop}{tidyselect variable indicating population ID}

\item{factrs}{character vector of rate-factors}

\item{df}{dataframe in which rows are populations, column 'factor_df' is a nested dataframe of rate-factors, and column 'pop_prods' is the rowProducts of 'factor_df'.}
}
\description{
Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object
}
\examples{
......
}
