% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgcc.R
\name{dgcc}
\alias{dgcc}
\title{Das Gupta equation 5.36 across N populations: Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.}
\usage{
dgcc(x, pop, id_vars, crossclassified)
}
\arguments{
\item{x}{dataframe consisting of one population, including variables indicating cross-classified structure, and a variable indicating size of each cell}

\item{pop}{variable name (character string) containing population identifier}

\item{id_vars}{character vector of variables indicating cross-classified structure.}

\item{crossclassified}{variable name (character string) containing cell sizes or proportions}
}
\value{
inputted data.frame is returned with the addition of variables for each of the the cross-classified variables representing the contribution to the population size.
}
\description{
Das Gupta equation 5.36 across N populations: Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.
}
