% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg2_1.R
\docType{data}
\name{dgeg2_1}
\alias{dgeg2_1}
\title{Das Gupta Example 2.1}
\format{
A data frame with 2 rows and 3 variables:
\describe{
\item{pop}{Population indicator}
\item{avg_earnings}{Mean earnings based on those who earned}
\item{earner_prop}{Proportion of persons who earned}
}
}
\source{
\href{https://www.census.gov/library/publications/1993/demo/p23-186.html}{Das Gupta, Prithwis (1993). Standardization and Decomposition of Rates: A User's Manual. U.S. Bureau of the Census, Current Population Reports, Series P23-186. U.S. Government Printing Office, Washington, DC.}
}
\usage{
dgeg2_1
}
\description{
Mean Earnings as the Product of Two Factors for Black Males and White Males, 18 Years and Over: United States, 1980

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented}
}
\examples{
data(dgeg2_1)
dgnpop(dgeg2_1,
      pop = "pop",
      factors = c("avg_earnings", "earner_prop")) |>
 dg_table()

}
\keyword{datasets}
