% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg2_4.R
\docType{data}
\name{dgeg2_4}
\alias{dgeg2_4}
\title{Das Gupta Example 2.4}
\format{
A data frame with 2 rows and 6 variables:
\describe{
\item{pop}{Population indicator}
\item{prop_m}{Index of proportion married}
\item{noncontr}{Index of noncontraception}
\item{abort}{Index of induced abortion}
\item{lact}{Index of lactational infecundability}
\item{fecund}{Total fecundity rate}
}
}
\source{
\href{https://www.census.gov/library/publications/1993/demo/p23-186.html}{Das Gupta, Prithwis (1993). Standardization and Decomposition of Rates: A User's Manual. U.S. Bureau of the Census, Current Population Reports, Series P23-186. U.S. Government Printing Office, Washington, DC.}
}
\usage{
dgeg2_4
}
\description{
Total Fertility Rate as the Product of Five Factors: South Korea, 1960 and 1970

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented}
}
\examples{
data(dgeg2_4)
dgnpop(dgeg2_4,
       pop = "pop",
       factors = c("prop_m", "noncontr", "abort", "lact", "fecund")) |>
  dg_table()

}
\keyword{datasets}
