% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg4_1.R
\docType{data}
\name{dgeg4_1}
\alias{dgeg4_1}
\title{Das Gupta Example 4.1}
\format{
A data frame with 18 rows and 4 variables:
\describe{
\item{age_group}{Age group: 1 = \verb{[10,15)}, 2 = \verb{[15,20)}, ..., 8 = \verb{[45,50)}, 9 = \verb{[50,55)}}
\item{pop}{Population indicator}
\item{Lx}{Stationary population values (5Lx)}
\item{mx}{Age-specific fertility rates for females}
}
}
\source{
"Standardization and decomposition of rates: A user’s manual", Das Gupta (1993)
}
\usage{
dgeg4_1
}
\description{
Female Intrinsic Growth Rate per Person as a Function of Two Vector-Factors: United States, 1960 and 1965

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented, but recognise that some of the terminology and classifications no longer align with current practices}
}
\examples{
data(dgeg4_1)
## rate function:
RF4.1 <- function(A, B) {
  idx <- seq_len(length(A))
  mu0 <- sum(A * B / 100000)
  mu1 <- sum((5 * idx + 7.5) * A * B / 100000)
  r1 <- log(mu0) * (mu0 / mu1)
  while (TRUE) {
    Nr1 <- 0
    Dr1 <- 0
    Nr1 <- Nr1 + sum(exp(-r1 * (5 * idx + 7.5)) * A * (B / 100000))
    Dr1 <- Dr1 - sum((5 * idx + 7.5) * exp(-r1 * (5 * idx + 7.5)) * A * (B / 100000))
    r2 <- r1 - ((Nr1 - 1) / Dr1)
    if (abs(r2 - r1) <= .0000001) {
      break
    }
    r1 <- r2
  }
  return(r2)
}

## crude rates:
RF4.1(A = dgeg4_1$Lx[1:9], B = dgeg4_1$mx[1:9])
RF4.1(A = dgeg4_1$Lx[10:18], B = dgeg4_1$mx[10:18])

## decomposition:
dgnpop(dgeg4_1,
       pop = "pop",
       factors = c("Lx", "mx"),
       id_vars = "age_group",
       ratefunction = "RF4.1(Lx,mx)") |>
  dg_table()

}
\keyword{datasets}
