% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg5_3.R
\docType{data}
\name{dgeg5_3}
\alias{dgeg5_3}
\title{Das Gupta Example 5.3}
\format{
A data frame with 44 rows and 5 variables:
\describe{
\item{race}{race group: 1 = White, 2 = non-White}
\item{age}{Age group: 1 = \verb{[0,1)}, 2 = \verb{[1,4]}, 3 = \verb{[5,14]}, 4 = \verb{[15,24]}, ..., 9 = \verb{[65,74]}, 10 = \verb{[75,84]}, 11 = \verb{[85,Inf)}}
\item{pop}{Population indicator}
\item{size}{Number of people in age-race-group i}
\item{rate}{Age-race-specific death rate (deaths per 1000 people)}
}
}
\source{
\href{https://www.census.gov/library/publications/1993/demo/p23-186.html}{Das Gupta, Prithwis (1993). Standardization and Decomposition of Rates: A User's Manual. U.S. Bureau of the Census, Current Population Reports, Series P23-186. U.S. Government Printing Office, Washington, DC.}
}
\usage{
dgeg5_3
}
\description{
Population (in thousands) and Death Rates (per 1,000 Population) by Age and Race: United States, 1970 and 1985

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented}
}
\examples{
data(dgeg5_3)

dgnpop(dgeg5_3,
       pop = "pop",
       factors = c("rate"),
       id_vars = c("race", "age"),
       crossclassified = "size") |>
  dg_table()

}
\keyword{datasets}
