% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg6_5.R
\docType{data}
\name{dgeg6_5}
\alias{dgeg6_5}
\title{Das Gupta Example 6.5}
\format{
A data frame with 30 rows and 6 variables:
\describe{
\item{pop}{Population indicator}
\item{agegroup}{Age group:\verb{[15,19]}, \verb{[20,24]}, ..., \verb{[35,39]}, \verb{[40,44]}}
\item{A}{Age-structure of women in childbearing ages (number of women in age group as a proportion of total women age 15-44)}
\item{B}{Marital status structure within childbearing age groups (proportion of women in age group i who are unmarried)}
\item{C}{Age-specific nonmarital fertility rate (proportion of unmarried women in age group i giving birth)}
\item{D}{Age-specific marital fertility rate (proportion of married women in age group i giving birth)}
}
}
\source{
\href{https://www.census.gov/library/publications/1993/demo/p23-186.html}{Das Gupta, Prithwis (1993). Standardization and Decomposition of Rates: A User's Manual. U.S. Bureau of the Census, Current Population Reports, Series P23-186. U.S. Government Printing Office, Washington, DC.}
}
\usage{
dgeg6_5
}
\description{
Illegitimacy Ratio as a Function of Four Vector-Factors: United States, White Population, 1963, 1968, 1973, 1978, and 1983.

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented}
}
\examples{
data(dgeg6_5)

dgnpop(dgeg6_5,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "1000*sum(A*B*C) / (sum(A*B*C) + sum(A*(1-B)*D))") |>
  dg_table()

dgnpop(dgeg6_5,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "1000*sum(A*B*C) / (sum(A*B*C) + sum(A*(1-B)*D))") |>
  dg_plot()

}
\keyword{datasets}
