% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgnpop.R
\name{dgnpop}
\alias{dgnpop}
\title{Prithwis Das Gupta's 1993 standardisation and decomposition of rates over K rate-factors and N populations.}
\usage{
dgnpop(
  x,
  pop,
  factors,
  id_vars = NULL,
  crossclassified = NULL,
  ratefunction = NULL,
  agg = TRUE,
  baseline = NULL,
  quietly = TRUE,
  diffs = FALSE
)
}
\arguments{
\item{x}{dataframe or tibble object, with columns specifying 1) population, 2) each rate-factor to be considered, and (optionally) 3) variables indicating underlying subpopulations}

\item{pop}{name (character string) of variable indicating population}

\item{factors}{names (character vector) of variables indicating compositional factors}

\item{id_vars}{character vector of variables indicating sub-populations}

\item{crossclassified}{character string of variable indicating size of sub-population. If specified, the proportion of each population in a given sub-population (e.g. each age-sex combination) is re-expressed as a product of symmetrical expressions representing the different variables (age, sex) constituting the sub-populations. These expressions are then used as compositional factors in the standardisation. If NULL, then providing a single variable as a compositional factor that represents the proportion of the population in each given sub-population will combine the contribution of all sub-population variables.}

\item{ratefunction}{user defined character string in R syntax that when evaluated specifies the function defining the rate as a function of factors. if NULL then will assume rate is the product of all factors. When sub-populations are provided, this should aggregate to a summary value (e.g., for the simple product rate this should be provided as \code{"sum(A*B*C*)"}.). User-defined functions can also be provided, as whatever string is given here will be parsed and evaluated as any other R code (see example eg4.4).}

\item{agg}{logical indicating whether, when cross-classified data is used, to output should be aggregated up to the population level}

\item{baseline}{baseline population to standardise against. if NULL then will do Das Gupta's full N-population standardisation.}

\item{quietly}{logical indicating whether interim messages should be outputted indicating progress through the K factors and N populations}

\item{diffs}{logical indicating whether to return list of standardised rates and rate-differences, or just the standardised rates.}
}
\value{
data.frame containing K-a standardised rates (or differences) for each population.
\itemize{
\item \code{rate}: standardised rate such that factor a is from population p and all other factors are averaged across populations, f(a^p,...)
\item \code{pop}: population p for which factor a is taken from
\item \code{std.set}: set of N populations (minus p) across which the standardisation has been performed
\item \code{factor}: name of factor a that is being considered, such that for the set of factors K, the \{K-a\}-standardised rate is returned
}
}
\description{
Prithwis Das Gupta's 1993 standardisation and decomposition of rates over K rate-factors and N populations.
}
\details{
Population rates are often composed of various different compositional factors. Standardisation techniques calculate the rate were a set of factors to be held constant (either with a specific population as standard, or at the average of the populations). Decomposition methods quantify the amount of the difference between two population crude rate that is due to differences in population characteristics.

Das Gupta's general solution for the decomposition of two rates can be written as:

\deqn{ \Delta\text{crude-}r = \sum\limits_{\vec{\alpha} \in K}Q(\vec{\alpha}^p) - Q(\vec{\alpha}^{p'}) }

Where \eqn{K} is the set of factors \eqn{\alpha, \beta, ..., \kappa}, which may take the form of vectors over sub-populations \eqn{i}. \eqn{Q(\vec{\alpha}^p)} denotes the rate in population \eqn{p} holding all factors other than \eqn{\alpha} --- \eqn{K \setminus \alpha} --- equal (standardised across populations \eqn{p} and \eqn{p'}). The total crude rate difference is the sum of all standardised-rate differences, and the standardisation \eqn{Q} is expressed as:

\deqn{Q(\vec{\alpha}^p) = \sum\limits_{j=1}^{\lfloor \frac{|K|}{2} \rfloor} \frac{ \sum\limits_{L \in {K \setminus \{\alpha\} \choose j-1}}f(\{L^p,(K\setminus L)^{p'},\vec{\alpha}^p\}) + f(\{L^{p'},(K\setminus L)^p,\vec{\alpha}^p\})} { |K| {|K| -1\choose j-1} } }

Where \eqn{f(K)} is the function that defines the calculation of the rate
}
\examples{
## 2 populations, R=ab
data(dgeg2_1)

dgnpop(dgeg2_1,
       pop = "pop",
       factors = c("avg_earnings", "earner_prop")) |>
  dg_table()

## 2 populations, R=abc
data(dgeg2_2)

dgnpop(dgeg2_2,
       pop = "pop",
       factors = c("birthsw1549", "propw1549", "propw")) |>
  dg_table()

## 2 populations, R=abcd
data(dgeg2_3)

dgnpop(dgeg2_3,
   pop = "pop",
   factors = c("birth_preg", "preg_actw", "actw_prop", "w_prop" )) |>
  dg_table()

## 2 populations, R=abcde
data(dgeg2_4)

dgnpop(dgeg2_4,
       pop = "pop",
       factors = c("prop_m", "noncontr", "abort", "lact", "fecund")) |>
  dg_table()

## 2 populations, vector factors, R=sum(abc)
data(dgeg4_3)

dgnpop(dgeg4_3,
       pop = "pop",
       factors = c("bm", "mw", "wp"),
       ratefunction = "sum(bm*mw*wp)") |>
 dg_table()

## 2 populations, R=f(ab)
data(dgeg3_1)

dgnpop(dgeg3_1,
       pop = "pop",
       factors = c("crude_birth", "crude_death"),
       ratefunction = "crude_birth-crude_death") |>
  dg_table()

## 2 populations, vector factors, R=f(abcd)
data(dgeg4_4)

dgnpop(dgeg4_4,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "sum(A*B*C) / (sum(A*B*C) + sum(A*(1-B)*D))") |>
  dg_table()

#### alternatively:
myratef <- function(a, b, c, d) {
 return(sum(a * b * c) / (sum(a * b * c) + sum(a * (1 - b) * d)))
}

dgnpop(dgeg4_4,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "myratef(A,B,C,D)") |>
  dg_table()

## 2 populations, 1 sub-population variable
data(dgeg5_1)

### Option 1:
dgnpop(dgeg5_1,
       pop = "pop",
       factors = c("size","rate"),
       ratefunction = "sum( (size/sum(size)) * rate )") |>
  dg_table()

### Option 2:
dgnpop(dgeg5_1,
       pop = "pop",
       factors = c("rate"),
       id_vars = "age_group",
       crossclassified = "size") |>
  dg_table()

## 2 cross-classified variables, 2 populations, R=sum(w*r)
data(dgeg5_3)

dgnpop(dgeg5_3,
       pop = "pop",
       factors = c("rate"),
       id_vars = c("race", "age"),
       crossclassified = "size") |>
  dg_table()

## 5 populations, R = f(abcd)
data(dgeg6_5)

dgres <- dgnpop(dgeg6_5,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "1000*sum(A*B*C) / (sum(A*B*C) + sum(A*(1-B)*D))")

dg_table(dgres)
dg_plot(dgres)

}
