% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Npops_factor_effects.R
\name{Npops_factor_effects}
\alias{Npops_factor_effects}
\title{The full Das Gupta 6.12 equation for a factor-effect between 2 populations
Outputs a tibble object specifying the factor-effect standardised across all populations}
\usage{
Npops_factor_effects(factoreffs, ps, all_p)
}
\arguments{
\item{factoreffs}{a dataframe/tibble object of ?? can't remember!!!}

\item{ps}{vector of length 2 specifying a possible pairwise comparison of populations}

\item{all_p}{character or numeric vector of all N populations}
}
\description{
The full Das Gupta 6.12 equation for a factor-effect between 2 populations
Outputs a tibble object specifying the factor-effect standardised across all populations
}
\examples{
......
}
