% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_popstr.R
\name{split_popstr}
\alias{split_popstr}
\title{Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.}
\usage{
split_popstr(df, id_vars, nvar)
}
\arguments{
\item{df}{dataframe consisting of one population, including variables indicating cross-classified structure, and a variable indicating size of each cell}

\item{id_vars}{character vector of variables indicating cross-classified structure.}

\item{nvar}{variable name (character string) containing cell sizes}
}
\description{
Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.
}
