% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DasGupt_Npop.R
\name{DasGupt_Npop}
\alias{DasGupt_Npop}
\title{Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object}
\usage{
DasGupt_Npop(df, pop, ..., baseline = NULL, id_vars = NULL,
  ratefunction = NULL)
}
\arguments{
\item{df}{dataframe or tibble object, with columns specifying 1) population and 2) each rate-factor to be considered}

\item{pop}{tidyselect variable indicating population ID}

\item{...}{tidyselect variables indicating rate-factors}
}
\description{
Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object
}
\examples{
......
}
