% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg354.R
\name{dg354}
\alias{dg354}
\title{Das Gupta equation 3.54. internal function called by dg2pop.}
\usage{
dg354(df2, i, pop, factors, id_vars, ratefunction, quietly = TRUE)
}
\arguments{
\item{df2}{list of 2 population dataframes, in which each one contains data for all factors for the relevant population, along with variables indicating population and subpopulations}

\item{i}{the index of the factors vector which is not being adjusted for (the alpha in "P-alpha standardised rates")}

\item{pop}{name (character string) of variable indicating population}

\item{factors}{names (character vector) of variables indicating compositional factors}

\item{id_vars}{character vector of variables indicating sub-populations}

\item{ratefunction}{user defined character string in R syntax that when evaluated specifies the function defining the rate as a function of factors. if NULL then will assume rate is the product of all factors.}

\item{quietly}{logical indicating whether interim messages should be outputted indicating progress}
}
\value{
data.frame object including K-a standardised rates for each population for given factor a, along with differences between standardised rates
}
\description{
Das Gupta equation 3.54. internal function called by dg2pop.
}
