% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg611.R
\name{dg611}
\alias{dg611}
\title{Das Gupta equation 6.11: Standardises rates across populations
Outputs a tibble object of the rate factor F in population Y, standardised across all populations}
\usage{
dg611(srates2, all_p, y, fctr)
}
\arguments{
\item{srates2}{a dataframe/tibble object of standardised rates from dg2pop}

\item{all_p}{character or numeric vector of all N populations}

\item{y}{character of numeric value indicating a single population}

\item{fctr}{string indicating rate-factor being standardised.}
}
\description{
Das Gupta equation 6.11: Standardises rates across populations
Outputs a tibble object of the rate factor F in population Y, standardised across all populations
}
\examples{
......
}
