% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg612.R
\name{dg612}
\alias{dg612}
\title{The full Das Gupta 6.12 equation for a factor-effect between 2 populations
Outputs a tibble object specifying the factor-effect standardised across all populations}
\usage{
dg612(factoreffs2, all_p, ps, fctr)
}
\arguments{
\item{all_p}{character or numeric vector of all N populations}

\item{ps}{vector of length 2 specifying a possible pairwise comparison of populations}

\item{factoreffs}{a dataframe/tibble object of ?? can't remember!!!}
}
\description{
The full Das Gupta 6.12 equation for a factor-effect between 2 populations
Outputs a tibble object specifying the factor-effect standardised across all populations
}
\examples{
......
}
