% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg612.R
\name{dg612}
\alias{dg612}
\title{Das Gupta equation 6.12 for a differences between 2 populations when standardised across N populations.}
\usage{
dg612(srates, all_p, ps, factor)
}
\arguments{
\item{srates}{a dataframe output from dg2p}

\item{all_p}{character or numeric vector of all N populations}

\item{ps}{vector of length 2 specifying a possible pairwise comparison of populations}

\item{factor}{character string indicating name of factor}
}
\value{
data.frame object including K-a standardised-rate-differences for each population for given factor a, across N populations
}
\description{
Das Gupta equation 6.12 for a differences between 2 populations when standardised across N populations.
}
