% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg612.R
\name{dg612}
\alias{dg612}
\title{The full Das Gupta 6.12 equation for a factor-effect between 2 populations when standardised across N populations.
Outputs a dataframe object specifying the factor-effect standardised across all populations}
\usage{
dg612(srates, all_p, ps, factor)
}
\arguments{
\item{srates}{a dataframe output from dg2p}

\item{all_p}{character or numeric vector of all N populations}

\item{ps}{vector of length 2 specifying a possible pairwise comparison of populations}

\item{factor}{character string indicating name of factor}
}
\description{
The full Das Gupta 6.12 equation for a factor-effect between 2 populations when standardised across N populations.
Outputs a dataframe object specifying the factor-effect standardised across all populations
}
