% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_plot.R
\name{dg_plot}
\alias{dg_plot}
\title{Creates a plot of Das Gupta adjusted rates across the set of populations}
\usage{
dg_plot(dgo)
}
\arguments{
\item{dgo}{DG output}
}
\description{
Creates a plot of Das Gupta adjusted rates across the set of populations
}
