% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_table.R
\name{dg_table}
\alias{dg_table}
\title{Creates a small table of Das Gupta extract adjusted rates & difference effects for 2 specified populations.}
\usage{
dg_table(dgo, pop1 = NULL, pop2 = NULL)
}
\arguments{
\item{dgo}{DG output}

\item{pop1}{population 1 (character/numeric)}

\item{pop2}{population 2 (character/numeric)}
}
\description{
Creates a small table of Das Gupta extract adjusted rates & difference effects for 2 specified populations.
}
