% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_table.R
\name{dg_table}
\alias{dg_table}
\title{Creates a small table of Das Gupta extract adjusted rates & difference effects for 2 specified populations.}
\usage{
dg_table(df, pop1, pop2)
}
\arguments{
\item{df}{DG output}

\item{pop1}{population 1 (string or numeric)}

\item{pop2}{population 2 (string or numeric)}
}
\description{
Creates a small table of Das Gupta extract adjusted rates & difference effects for 2 specified populations.
}
