% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGadjust_ratefactor.R
\name{DGadjust_ratefactor}
\alias{DGadjust_ratefactor}
\title{Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object}
\usage{
DGadjust_ratefactor(df2, pop, i, factrs, ratefunction)
}
\arguments{
\item{df2}{dataframe in which rows are populations, column 'factor_df' is a nested dataframe of rate-factors, and column 'pop_prods' is the rowProducts of 'factor_df'.}

\item{pop}{tidyselect variable indicating population ID}

\item{i}{the index of the factrs vector which is being adjusted for (the alpha in P-alpha)}

\item{factrs}{character vector of rate-factors}

\item{ratefunction}{allows user to define rate as a specific function F of factors. This should be a character string of the r syntax, with the factor names. Defaults to the product (e.g., "a*b").}
}
\description{
Inner function called by DasGupt_2pop and DasGupt_Npop
Outputs a ? object
}
\examples{
......
}
