% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg4_4.R
\docType{data}
\name{dgeg4_4}
\alias{dgeg4_4}
\title{Das Gupta Example 4.4}
\format{
A data frame with 12 rows and 6 variables:
\describe{
\item{pop}{Population indicator}
\item{agegroup}{Age group: 1 = \verb{[15,19]}, 2 = \verb{[20,24]}, ..., 5 = \verb{[35,39]}, 6 = \verb{[40,44]}}
\item{A}{Age-structure of women in childbearing ages (number of women in age group as a proportion of total women age 15-44)}
\item{B}{Marital status structure within childbearing age groups (proportion of women in age group i who are unmarried)}
\item{C}{Age-specific nonmarital fertility rate (proportion of unmarried women in age group i giving birth)}
\item{D}{Age-specific marital fertility rate (proportion of married women in age group i giving birth)}
}
}
\source{
"Standardization and decomposition of rates: A user’s manual", Das Gupta (1993)
}
\usage{
dgeg4_4
}
\description{
Illegitimacy Ratio as a Function of Four Vector-Factors: United States, White Population, 1963 and 1983

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented, but recognise that some of the terminology and classifications no longer align with current practices}
}
\examples{
data(dgeg4_4)

## Option 1:
dgnpop(dgeg4_4,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "sum(A*B*C) / (sum(A*B*C) + sum(A*(1-B)*D))") |>
  dg_table()

## alternatively, define a separate rate function:
myratef <- function(a, b, c, d) {
  return(sum(a * b * c) / (sum(a * b * c) + sum(a * (1 - b) * d)))
}

dgnpop(dgeg4_4,
       pop = "pop",
       factors = c("A", "B", "C", "D"),
       id_vars = "agegroup",
       ratefunction = "myratef(A,B,C,D)") |>
  dg_table()

}
\keyword{datasets}
