% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeg6_6.R
\docType{data}
\name{dgeg6_6}
\alias{dgeg6_6}
\title{Das Gupta Example 6.6: US Population Birth Rates}
\format{
A data frame with 459 rows and 4 variables:
\describe{
\item{year}{Population indicator}
\item{agebin}{Age-sex group: \code{F [10,14]}, \code{F [15,19]}, \code{F [20,24]}, \code{F [25,29]}, \code{F [30,34]}, \code{F [35,39]}, \code{F [40,44]}, \code{F [45,49]}, \verb{Remainder (F [0,10), F [50, Inf), M [0,Inf)})}
\item{thous}{Population (in thousands) in age-sex group i}
\item{birthrate}{Age-sex-specific birth rates (per 1000 population)}
}
}
\source{
"Standardization and decomposition of rates: A user’s manual", Das Gupta (1993)
}
\usage{
dgeg6_6
}
\description{
Population and Birth Rates by Nine Age-Sex Groups: United States, 1940 to 1990

\emph{To allow for direct comparison with Das Gupta's original publication we have preserved these datasets as they were originally presented}
}
\examples{
data(dgeg6_6)

dgo_us <- dgnpop(dgeg6_6,
                 pop = "year",
                 factors = c("birthrate"),
                 id_vars = "agebin",
                 crossclassified = "thous")

dg_plot(dgo_us)

# Recreating Figure 1 (p.112) from Das Gupta 1993
# crude rates
plot(x = 1940:1990,
     y = by(dgeg6_6, dgeg6_6$year,
            \(x) sum(x$birthrate * (x$thous/sum(x$thous)))),
     type = "l", ylim = c(13,33))
# rates with 1940 as standard
baseline = dgeg6_6[dgeg6_6$year == 1940, ]
lines(x = 1940:1990,
      y = by(dgeg6_6, dgeg6_6$year,
             \(x) sum(x$birthrate * (baseline$thous/sum(baseline$thous)))),
      lty = "dotted")
# rates with 1990 as standard
baseline = dgeg6_6[dgeg6_6$year == 1990, ]
lines(x = 1940:1990,
      y = by(dgeg6_6, dgeg6_6$year,
             \(x) sum(x$birthrate * (baseline$thous/sum(baseline$thous)))),
      lty = "dashed")
# rates standardised with Das Gupta's method
lines(x = 1940:1990,
      y = dgo_us$rate[dgo_us$factor=="birthrate"],
      lty = "longdash", col = "red")

}
\keyword{datasets}
