% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgnpop.R
\name{dgnpop}
\alias{dgnpop}
\title{Prithwis Das Gupta's 1993 standardisation and decomposition of rates over P rate-factors and N populations.}
\usage{
dgnpop(
  x,
  pop,
  factors,
  id_vars = NULL,
  crossclassified = NULL,
  ratefunction = NULL,
  agg = TRUE,
  baseline = NULL,
  quietly = TRUE
)
}
\arguments{
\item{x}{dataframe or tibble object, with columns specifying 1) population, 2) each rate-factor to be considered, and (optionally) 3) variables indicating underlying subpopulations}

\item{pop}{name (character string) of variable indicating population}

\item{factors}{names (character vector) of variables indicating compositional factors}

\item{id_vars}{character vector of variables indicating sub-populations}

\item{crossclassified}{character string of variable indicating size of sub-population. If specified, the proportion of each population in a given sub-population (e.g. each age-sex combination) is re-expressed as a product of symmetrical expressions representing the different variables (age, sex) constituting the sub-populations. These expressions are then used as compositional factors in the standardisation. If NULL, then providing a single variable as a compositional factor that represents the proportion of the population in each given sub-population will combine the contribution of all sub-population variables.}

\item{ratefunction}{user defined character string in R syntax that when evaluated specifies the function defining the rate as a function of factors. if NULL then will assume rate is the product of all factors. When sub-populations are provided, this should aggregate to a summary value (e.g., for the simple product rate this should be provided as `"sum(A*B*C*)"`.). User-defined functions can also be provided, as whatever string is given here will be parsed and evaluated as any other R code (see example eg4.4).}

\item{agg}{logical indicating whether, when cross-classified data is used, to output should be aggregated up to the population level}

\item{baseline}{baseline population to standardise against. if NULL then will do Das Gupta's full N-population standardisation.}

\item{quietly}{logical indicating whether interim messages should be outputted indicating progress through the P factors and N populations}
}
\value{
data.frame with TODO finish this
}
\description{
Population rates are often composed of various different compositional factors. For example, Das Gupta 1992 expressed the birth rate as the product of the fertility rate, proportion of women who are of childbearing ages, and the proportion of women in the population. Comparisons of crude rates between populations can be misleading as different compositional factors may be contributing to different extents on any observed differences in the crude rate. Standardisation of rates ... TODO finish this description
}
\examples{
dgnpop()
}
