% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgnpop.R
\name{dgnpop}
\alias{dgnpop}
\title{Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object}
\usage{
dgnpop(
  df,
  pop,
  ...,
  baseline = NULL,
  id_vars = NULL,
  ratefunction = NULL,
  quietly = TRUE
)
}
\arguments{
\item{df}{dataframe or tibble object, with columns specifying 1) population and 2) each rate-factor to be considered}

\item{pop}{tidyselect variable indicating population ID}

\item{...}{tidyselect variables indicating rate-factors}

\item{baseline}{baseline population to standardise against. if NULL then will do Das Gupta's full N-population standardisation.}

\item{id_vars}{character vector of variable names in dataframe to attach back to output (useful if analysing sub-population specific rates)}

\item{ratefunction}{allows user to define rate as a specific function F of factors. This should be a character string of the r syntax, with the factor names. Defaults to the product (e.g., "a*b").}
}
\description{
Standardisation and decomposition of rates over P rate-factors and N populations.
Outputs a ? object
}
\examples{

# The case of 2 factors (2 populations)
eg2.1 <- data.frame(
  pop = c("black","white"),
  avg_earnings = c(10930, 16591),
  earner_prop = c(.717892, .825974)
)
dgnpop(eg2.1, pop=pop, avg_earnings, earner_prop)

# The case of 3 factors (2 populations)
eg2.3 <- data.frame(
  pop = c("austria","chile"),
  birthsw1549 = c(51.78746, 84.90502),
  propw1549 = c(.45919, .75756),
  propw = c(.52638, .51065)
)
dgnpop(eg2.3, pop=pop, birthsw1549, propw1549, propw)

# The case of 4 factors (2 populations)
eg2.5 <- data.frame(
  pop = c(1971, 1979),
  birth_preg = c(25.3, 32.7),
  preg_actw = c(.214, .290),
  actw_prop = c(.279, .473),
  w_prop = c(.949, .986)
)
dgnpop(eg2.5, pop=pop, birth_preg, preg_actw, actw_prop, w_prop)

# The case of 5 factors (2 populations)
eg2.7 <- data.frame(
  pop = c(1970, 1980),
  prop_m = c(.58, .72),
  noncontr = c(.76, .97),
  abort = c(.84, .97),
  lact = c(.66, .56),
  fecund = c(16.573, 16.158)
)

dgnpop(eg2.7, pop=pop, prop_m, noncontr, abort, lact, fecund)

# The case of 3 vector factors (2 populations)
eg4.5 <- data.frame(
  agegroup = rep(1:7, 2),
  pop = rep(c(1970, 1960), e = 7),
  bm = c(488, 452, 338, 156, 63, 22, 3,
         393, 407, 369, 274, 184, 90, 16),
  mw = c(.082, .527, .866, .941, .942, .923, .876,
         .122, .622, .903, .930, .916, .873, .800),
  wp = c(.056, .038, .032, .030, .026, .023, .019,
         .043, .041, .036, .032, .026, .020, .018)
)
dgnpop(eg4.5, pop=pop, bm, mw, wp, id_vars=c("agegroup"))
dgout <- dgnpop(eg4.5, pop=pop, bm, mw, wp, id_vars=c("agegroup"))
dg_rates(dgout)
}
