% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconv.R
\docType{data}
\name{reconv}
\alias{reconv}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(reconv)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(reconv)
dg_data <- DasGupt_Npop(reconv,pop=year,prev,age_str,id_vars=c(Age,Gender))

}
\references{
Scottish Government Reconviction data:
(\href{https://www.gov.scot/publications/reconviction-rates-scotland-2016-17-offender-cohort/}{2016/17})
}
\keyword{datasets}
