% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconv.R
\docType{data}
\name{reconv}
\alias{reconv}
\title{Scottish Reconvictions data 2004-2016}
\format{
An object of class \code{data.frame} with 130 rows and 8 columns.
}
\usage{
data(reconv)
}
\description{
Scottish Reconvictions data 2004-2016
}
\examples{
data(reconv)
dgnpop(reconv, pop="year", factors=c("prev_rate"), id_vars=c("Sex","Age"), crossclassified="offenders")$rates |> dg_plot()

}
\references{
Scottish Government Reconviction data:
(\href{https://www.gov.scot/publications/reconviction-rates-scotland-2016-17-offender-cohort/}{2016/17})
}
\keyword{datasets}
