% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_popstr.R
\name{split_popstr}
\alias{split_popstr}
\title{Das Gupta equation 5.36 for a single population: Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.}
\usage{
split_popstr(x, id_vars, nvar)
}
\arguments{
\item{x}{dataframe consisting of one population, including variables indicating cross-classified structure, and a variable indicating size of each cell}

\item{id_vars}{character vector of variables indicating cross-classified structure.}

\item{nvar}{variable name (character string) containing cell sizes}
}
\value{
inputted data.frame is returned with the addition of variables for each of the the cross-classified variables representing the contribution to the population size.
}
\description{
Das Gupta equation 5.36 for a single population: Decomposes cross-classified population structures into a set of symmetric proportions indicating contribution of individual structural variables.
}
