% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspop.R
\docType{data}
\name{uspop}
\alias{uspop}
\title{US population data 1940-1990}
\format{
An object of class \code{data.frame} with 459 rows and 4 columns.
}
\usage{
data(uspop)
}
\description{
US population data 1940-1990
}
\examples{
data(uspop)
dgnpop(uspop, pop = "year", factors = c("birthrate"), id_vars = "agebin", crossclassified = "thous") |>
dg_plot()

}
\keyword{datasets}
